Update Reconnexion Bluetooth but test KO
The reconnexion function is working: when no notification is received after retry delay the WakeUpDeviceAsync is called in GlucoseService : ok
But an error occurs during the receiving data when the connexion process is finished.
The error is : "No compatible code running. The selected debug engine does not support any code executing on the current thead (e.g. only native runtime code is executing)."
Hypothesis : an async process is presenting an error. To be studied after reading the code @aobah2204